requirements/base.txt
asgiref==3.8.1
Django==5.1.2
django-environ==0.11.2
djangorestframework==3.15.2
pillow==11.0.0
psycopg2==2.9.10
sqlparse==0.5.1
tzdata==2024.2
